Integrating redux
