+++ src/Actions/index.js
@@ -1,0 +1,2 @@
+export * from './ShoppingCartActions'
+export * from './HomeActions'
