+++ src/Actions/types.js
@@ -1,0 +1,4 @@
+export const ADD_TO_CART = 'add_to_cart'
+export const HIDE_ORDER = 'hide_order'
+export const DISPLAY_ORDER = 'display_order'
+export const SET_REVIEWS = 'set_reviews'
