+++ src/App.js
@@ -1,7 +1,9 @@
 import React, { Component } from 'react'
 import {BrowserRouter as Router, Route} from 'react-router-dom'
 
-import { Home, Menu } from './pages';
+
+import Home from './pages/Home'
+import Menu from './pages/Menu'
 
 class App extends Component {
   constructor(props) {
