+++ src/components/index.js
@@ -5,4 +5,3 @@ export * from './Featured'
 export * from './Reviews'
 export * from './Footer'
 export * from './MenuCompleto'
-export * from './ShoppingCart'
