--- src/pages/index.js
@@ -1,2 +1,0 @@
-export * from './Home'
-export * from './Menu'
